Implements the LabelCollapse inference method.
This method will build a auxiliary graphical model where most of the
labels are collapsed. Another proxy inference method will solve this new
auxiliary problem. If the labeling includes collapsed labels, some
labels will be uncollapsed and the process is repeated iteratively.

This approach should reduce the memory consumption of the underlying
proxy inference method while returning the same solution.

The implementation is meant to be used with the existing CombiLP
inference algorithm. Nevertheless it should work with all inference
methods which respect the inference interface.