Implements smarter logic for LabelCollapse algorithm.
The LabelCollapse inference method will now only introduce a new label
for “collapsed labels” if really necessary, i.e. if there exist
collapsed label for the respective variable. This will lower the memory
requirements (for the proxy solver) and also shortens the runtime of the
inference method (reduces number of necessary iterations).