Implements a unit test for the LabelCollapse inference.
The introduce tests will use a high number of labels, because this is
the primary use case for the new inference method. The test suite
bruteforces the optimal solution, so we need keep the problem size small
(large number of labels!).

Currently it uses the bruteforce algorithm as proxy inference method and
all tests should calculate the optimal energy. For a second pass the
CombiLP method is used as proxy.

The test suite should later be extended to test larger problems with
CombiLP and CPLEX.