Fixes factor assignment in LabelCollapse algorithm.
This commit introduces a check which prevents that no factor is found
for a given variable. If there is no such factor, no factor will be
updated (previously factor 0 was updated in such a case).

If no factors exists for a variable, no label will get collapsed. We do
only store epsilon values for factors and in the previous version all
labels got collapsed. This completely broke the algorithm, because there
was no way to uncollapse the labels.

The new behavior is also quite logical: If there are no factors, we
can’t distinguish the labels. So if one labels gets uncollapsed, all
labels will get uncollapsed (have same value). Thus, it makes sense to
uncollapse all labels by default.