Fixes returned labeling for the LabelCollapse inference.
The proxy solver is used to solve the generated auxiliary task.
Obviously, the labeling is not the same as the labeling for the original
problem. However, there exists a bijection for valid labeling of the
auxiliary and original labelings.

This commit correctly calculates the labeling for the original task. The
previous implementation was broken and returned the labeling of the
auxiliary task.