Optimizes run time of LabelCollapse inference method.
This commit changes the behavior of selecting collapsed and uncollapsed
labels for a given node. The new behavior is to select all labels where
the value is strictly smaller than the current epsilon value.

The initialization (reset) now needs to perform multiple passes of
epsilon calculations:

(1) Initialize to inverse neutral element.

(2) Calculate new epsilon for each factor.

(3) Calculate new (the next best) epsilon for each variable.

The new behavior speeds up the algorithm, because it needs to calculate
less iterations. The old version had to perform a lot of unnecessary
iterations (epsilon push and inference of auxiliary problem) after the
initialization of the inference method.