Sets LabelCollapse inferred value during last step.
The iterations of the LabelCollapse inference method calculate a lower
bound for the given problem. The previous implementation just passed
through the value und bound of the proxy inference (which is wrong).

This commits fixes the faulty behaviour: It updates the bound at each
step of the iteration and only on convergence sets the final inferred
value.