LabelCollapse pushes only unary epsilons.
This commit introduces a performance optimization for special kinds of
reparametrizations (benefits combination with CombiLP). If for other
factors than unary the factors the values are most of the time equal,
the new implementation will introduce less auxiliary labels.

Only the unary epsilons are pushed. The epsilons for other factors are
simply calculated (minimal value for collapsed labels). This simplifies
the whole alogrithm a lot.