Implements the possibility to “warm start” the CPLEX solver.
Warm starting is achieved by supplying a MIP starting point. The user
code just needs to supply an MRF labeling and then the CPLEX wrapper
will calculate the value for all the MIP variables and use it as an MIP
starting point.

This improves performance we solve related or similiar models sequentially.

(An even better implementation would just modify the existing CPLEX
environment by simply introducing new constraints and variables. Then
CPLEX is able to automatically update its environment by itself. The
only problem is that such an implementation is not easy to integrate
into the current OpenGM code base.)