Changes the internal representation for Unary-LabelCollapse.
The current implementation of LabelCollapse only considers unary
potentials and therefore a better internal representation is introduced.

The new implementations maintains a stack of collapsed labels for each
of the variables. An uncollapse step just needs no to pop one element
from the stack and include in the new graphical model.