Fixes a regression introduced by too heavy refactoring.
The problem is that we need to share the state of TRWSi solver between
the LP and ILP phase of CombiLP. In the second phase only needs a
reparametrizer, but the reparameterization framework is highly coupled
with the TRWSi solver.

The simplest solution for now is to just keep the TRWSi solver and
reparametrizer as class members. See comment in class definition of
CombiLP for some more details.