Fixes regression introduced by the process of refactoring.
The initial labeling for CombiLP is not the labeling returned by the
TRWSi inference solver, but a more special labeling related to
strict-arc-consistency.

Also there was a typo in the calculation of the boundary mask.

The refactored version updates value_ and bound_ inplace. Therefore we
also need to check the result of the visitor and act accordingly
(returning CONVERGENCE if requested).