Implements an extension for LabelCollapse.
It is not necessary to assign to all collapsed transistions between
variables the same potential.

Instead can fix all non-collapsed labels and check for the best local
transistions of the remaining collapsed labels.

This makes the LabelCollapse more efficient in general.

This optimization is implemented in the EpsilonFunction directly, so all
versions of LabelCollapse benefit from this optimization.