Implements epsilon expansion for LabelCollapse.
We choose an ε, such that all uncollapsed unaries are smaller than this
ε and all collapsed labels are greater than this ε.

To be able to find such an ε, we first need to adjust the label shape of
the problem. We achieve this by uncollapsing variables.

The idea is that we skip many unnecessary uncollapsing steps.
Additionally we could be also able to introduce some nice other
properties.