Fixes the newly added extension to LabelCollapse.
The old implementation of the extension was faulty.

We can’t rely on the precalculated epsilon values. Also the whole
process is a litte bit more complicated.

The current working implementation just calculates the potentials for
the EpsilonFunction (part of the Auxiliary GM) on-the-fly directly. Thus
the implementation doesn’t make any sense for the non-unary version of
LabelCollapse.

We should remove the generic code and just make the unary model builder
the default implementation.