Add FusionMove population to LabelCollapse.
The caller can provide a reasonable good labeling (for example an
approximate labeling). This labeling is now repeatedly checked against a
labeling which only consists of auxiliary labeling. For checking
a FusionMove is used.

FusionMove is guaranteed to only improve the labeling by moving labels
of one proposal to the other one. If the resulting new labeling contains
auxiliary labels, they get uncollapsed.

The idea is to speed up the initial uncollapsing phase by not using a
slow ILP solver. FusionMoves (which themselves use QPBO or similiar
algorithms) are much faster.